+++ test_allianceutils/tests/viewset_permissions/tests.py
@@ -11,10 +11,16 @@ from django.test import modify_settings
 from django.test import override_settings
 from django.test import TestCase
 from django.urls import reverse
+from rest_framework import viewsets
 from rest_framework.test import APIClient
+from rest_framework.test import APIRequestFactory
+from rest_framework.test import force_authenticate
 
+from allianceutils.api.permissions import GenericDjangoViewsetPermissions
 from test_allianceutils.tests.profile_auth.models import User
 from test_allianceutils.tests.viewset_permissions.models import NinjaTurtleModel
+from test_allianceutils.tests.viewset_permissions.models import SenseiRatModel
+from test_allianceutils.tests.viewset_permissions.views import NinjaTurtleSerializer
 
 USER_EMAIL = "test@example.com"
 USER_PASS = "password"
@@ -136,3 +142,64 @@ class ViewsetPermissionsTestCase(TestCase):
         test_methods("change_ninjaturtlemodel", {"change"})
 
         test_methods("delete_ninjaturtlemodel", {"delete"})
+
+
+    @modify_settings(
+        AUTHENTICATION_BACKENDS={
+            "append": 'test_allianceutils.tests.viewset_permissions.tests.IgnoreObjectsBackend',
+        }
+    )
+    def test_get_model_permission_viewset(self):
+        factory = APIRequestFactory()
+
+        # These two ViewSets are equivalent; they just use different hooks to
+        # return the model to use for permission checks
+        class NinjaTurtleGetModelPermission1(GenericDjangoViewsetPermissions):
+            def get_model(self, view):
+                return SenseiRatModel
+
+        class NinjaTurtleGetModelViewSet1(viewsets.ModelViewSet):
+            queryset = NinjaTurtleModel.objects.all()
+            serializer_class = NinjaTurtleSerializer
+
+            permission_classes = [NinjaTurtleGetModelPermission1]
+
+        class NinjaTurtleGetModelViewSet2(viewsets.ModelViewSet):
+            queryset = NinjaTurtleModel.objects.all()
+            serializer_class = NinjaTurtleSerializer
+
+            permission_classes = [GenericDjangoViewsetPermissions]
+
+            def get_permission_model(self):
+                return SenseiRatModel
+
+
+        def test_methods(with_grant_permission: Optional[str] = None, grant_model=NinjaTurtleModel,
+                should_succeed: Set = set()):
+            turtle_id = self.turtle.id
+            for viewset in (NinjaTurtleGetModelViewSet1, NinjaTurtleGetModelViewSet2):
+                with self.subTest(f"{viewset.__name__} with_grant_permission={with_grant_permission} grant_model={grant_model.__name__} should_success={should_succeed}"):
+                    with transaction.atomic():
+                        if with_grant_permission:
+                            self.grant_permission(with_grant_permission, grant_model)
+                        view = viewset.as_view({"get": "list"})
+                        # Refetch user... something was caching permissions
+                        user = User.objects.get(pk=self.user.pk)
+                        request = factory.get("")
+                        force_authenticate(request, user=user)
+                        response = view(request).render()
+                        self.assertEqual(response.status_code, 200 if "view" in should_succeed else 403)
+
+                        view = viewset.as_view({"get": "retrieve"})
+                        request = factory.get("")
+                        force_authenticate(request, user=user)
+                        response = view(request, pk=turtle_id).render()
+                        self.assertEqual(response.status_code, 200 if "view" in should_succeed else 403)
+
+                        transaction.set_rollback(True)
+
+        test_methods()
+        test_methods("view_ninjaturtlemodel")  # should fail; perm lookup would hit SenseiRat who is going to say no
+        test_methods("view_senseiratmodel", SenseiRatModel, {"view"})  # should succeed with permission from SenseiRat
+
+